Error definitions are separated from sdk package
