+++ packages/identity-cardinal/src/identity-cardinal-twitter.error.ts
@@ -1,0 +1,9 @@
+import { IdentityError } from '@dialectlabs/sdk';
+
+export class CardinalIdentityError extends IdentityError {
+  static ignoreMatcher = ['Account does not exist']; // warning, not error
+
+  constructor(msg?: string) {
+    super(CardinalIdentityError.name, 'Cardinal Identity Error', msg);
+  }
+}
